+++ src/settings/MappingProfiles/detailsSections/hooks/index.js
@@ -1,1 +1,5 @@
+export * from './useFieldMappingBoolFieldValue';
+export * from './useFieldMappingFieldValue';
+export * from './useFieldMappingRefValues';
+export * from './useFieldMappingValueFromLookup';
 export * from './useOrganizationValue';